requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
python-jose
argon2